pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-explore"
version = "0.1.0"
description = "Flux API"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-multipart>=0.0.5",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=1.4.23",
    "alembic>=1.7.1",
    "psycopg2-binary>=2.9.1",
    "pydantic>=1.8.2",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=0.19.0",
    "torch>=2.2.0",
    "diffusers>=0.24.0",
    "transformers>=4.36.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

linux = [
    "xformers>=0.0.22.post7",
]

gguf = [
    "llama-cpp-python>=0.2.0",
    "pillow>=9.5.0",
    "numpy>=1.24.0",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.uv]
# Default PyPI index
[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true 